@@ -27,7 +27,7 @@ describe LogsController do
   describe "GET edit" do
     it "should assign project to @project" do
       log = FactoryGirl.create(:log, :user => @user, :firm => @user.firm)
-      get :edit, :id => log
+      get :edit, :id => log, :format => 'js'
       assigns(:log).should eq(log) 
     end 
   end
@@ -35,12 +35,12 @@ describe LogsController do
     context "with valid attributes" do
       it "creates a new log" do
         expect{
-          post :create, log: FactoryGirl.attributes_for(:log)
+          post :create, log: FactoryGirl.attributes_for(:log), :format => 'js'
         }.to change(Log,:count).by(1)
       end
       
       it "gives flash notice when creating new log" do
-        post :create, log: FactoryGirl.attributes_for(:log)
+        post :create, log: FactoryGirl.attributes_for(:log), :format => 'js'
         flash[:notice].should_not be_nil 
       end
     end 
@@ -54,7 +54,7 @@ describe LogsController do
     
   
     it "changes @log's attributes" do
-      put :update, id: @log, log: FactoryGirl.attributes_for(:log, :event => "something else")
+      put :update, id: @log, log: FactoryGirl.attributes_for(:log, :event => "something else"), :format => 'js'
       @log.reload
       @log.event.should eq("something else")
     end
@@ -69,14 +69,14 @@ end
     
     it "deletes the contact" do
       expect{
-        delete :destroy, id: @log        
+        delete :destroy, id: @log , :format => 'js'       
       }.to change(Log,:count).by(-1)
     end
   end
   describe 'post start tracking' do
     it "creates a new log" do
         expect{
-          post :start_tracking, log: FactoryGirl.attributes_for(:log)
+          post :start_tracking, log: FactoryGirl.attributes_for(:log), :format => 'js'
         }.to change(Log,:count).by(1)
       end
     it 'should get vaild repsonse' do 
@@ -89,7 +89,7 @@ end
   
     context "valid attributes" do
       it "changes @log's attributes" do
-        put :stop_tracking, id: @log, log: FactoryGirl.attributes_for(:log, :event => "something else")
+        put :stop_tracking, id: @log, log: FactoryGirl.attributes_for(:log, :event => "something else"), :format => 'js'
         assigns(:log).should eq(@log)
         @log.reload
         @log.event.should eq("something else")
@@ -102,7 +102,7 @@ end
       project.users << @user
       todo = FactoryGirl.create(:todo, user: @user, firm: @user.firm, project: project)
       log = FactoryGirl.create(:log, todo: todo, project: project, user: @user, firm: @user.firm)
-      get :get_logs_todo, todo_id: todo
+      get :get_logs_todo, todo_id: todo, :format => 'js'
       assigns(:todo).should eq(todo)
     end
     it 'should send it off to LogWorker' do
