@@ -7,7 +7,6 @@ class MilestonesController < ApplicationController
         flash[:notice] = flash_helper('Milestone was successfully created.')
         format.js { render :action => "create_success"}
         else
-        format.xml  { render :xml => @log.errors, :status => :unprocessable_entity }
         format.js { render :action => "failure"}
       end
      end
@@ -23,7 +22,7 @@ class MilestonesController < ApplicationController
       end
     else
       respond_to do |format|
-        format.json { render :nothing =>  true }
+        format.js
       end
     end
   end
