@@ -6,9 +6,9 @@ Squadlink::Application.configure do
   # your test database is "scratch space" for the test suite and is wiped
   # and recreated between test runs.  Don't rely on the data there!
   config.cache_classes = true
-
+  config.eager_load = false
   # Log error messages when you accidentally call methods on nil.
-  config.whiny_nils = true
+  
 
   # Show full error reports and disable caching
   config.consider_all_requests_local       = true
@@ -39,5 +39,4 @@ Squadlink::Application.configure do
   
 
   # Raise exception on mass assignment protection for Active Record models
-  config.active_record.mass_assignment_sanitizer = :strict
 end
