@@ -4,4 +4,5 @@
 # If you change this key, all old signed cookies will become invalid!
 # Make sure the secret is at least 30 characters and all random,
 # no regular words or you'll be exposed to dictionary attacks.
-Squadlink::Application.config.secret_token = '1979188179ca453640f232bf1e5f40892ff63d11b9c92f5013a51d4d474d9bfac23dd8b057d833aa149163dd5dd10d493dd8ed58a8f5d552e68aade58f111b86'
+# Squadlink::Application.config.secret_token = '1979188179ca453640f232bf1e5f40892ff63d11b9c92f5013a51d4d474d9bfac23dd8b057d833aa149163dd5dd10d493dd8ed58a8f5d552e68aade58f111b86'
+Squadlink::Application.config.secret_key_base = '787ef0ced9066789754691b8555cc67c994b5d04326f53aef0422d24ae41ef042440480adb00c97380b2f620b8a45b4156538eb6befacfce740c44da3d0a581a'
