@@ -12,7 +12,7 @@ describe TabsController do
 	 	let(:milestone)	{FactoryGirl.create(:milestone, project: project, firm: project.firm)}
 
 	  it "should get milestones" do
-	  	get :tabs_milestones, id: project.id, class: "Project"
+	  	get :tabs_milestones, id: project.id, class: "Project", :format => 'js'
 	  	assigns(:milestones).should == [milestone]
 	  end
 	end	
@@ -22,7 +22,7 @@ describe TabsController do
 	 	let(:todo2)			{FactoryGirl.create(:todo, project: project, firm: @user.firm, user: @user, completed: true, due: Date.today)}
 
 	  it "should get todos" do
-	  	get :tabs_todos, id: project.id, class: "Project"
+	  	get :tabs_todos, id: project.id, class: "Project", :format => 'js'
 	  	assigns(:not_done_todos).should == [todo1]
 	  	assigns(:done_todos).should == [todo2]
 	  	assigns(:members).should == [@user]
@@ -33,7 +33,7 @@ describe TabsController do
 	 	let(:log)	{FactoryGirl.create(:log, project: project, firm: @user.firm, user: @user, log_date: Date.today)}
 
 	  it "should get logs" do
-	  	get :tabs_logs, id: project.id, class: "Project"
+	  	get :tabs_logs, id: project.id, class: "Project", :format => 'js'
 	  	assigns(:logs).should == [log]
 	  end
 	end	
@@ -42,7 +42,7 @@ describe TabsController do
 
 	 	it "should get users" do
 	  	project.users << @user
-	  	get :tabs_users, id: project.id, class: "Project"
+	  	get :tabs_users, id: project.id, class: "Project", :format => 'js'
 	  	assigns(:users).should == [@user]
 	  end
 	end	
@@ -51,7 +51,7 @@ describe TabsController do
 	 	let(:log)	{FactoryGirl.create(:log, project: project, firm: @user.firm, user: @user, log_date: Date.today)}
 
 	  it "should get statistics" do
-	  	get :tabs_statistics, id: project.id, class: "Project"
+	  	get :tabs_statistics, id: project.id, class: "Project", :format => 'js'
 	  	assigns(:logs).should == [log]
 	  end
 	end	
@@ -61,7 +61,7 @@ describe TabsController do
 
 	  it "should get statistics" do
 	  	project.users << @user
-	  	get :tabs_spendings, id: project.id, class: "Project"
+	  	get :tabs_spendings, id: project.id, class: "Project", :format => 'js'
 	  	assigns(:logs).should == [log]
 	  	assigns(:users).should == [@user]
 	  end
